Adjust `DataTypes.sum` to use `long` array, addressing potential integer overflow issue

+++ src/DataTypes.java
@@ -1,9 +1,9 @@
 public class DataTypes {
     // TODO TASK 1: fix this code so that it passes the test in DataTypesTest.java -> DONE
-    public static long sum(int[] numbers) {
+    public static long sum(long[] numbers) {
         long s = 0; // variable to accumulate the sum in!// changed to long to avoid integer overflow.
         // below is a "foreach" loop which iterates through numbers
-        for (int x : numbers) {
+        for (long x : numbers) {
             s += x;
         }
         return s;
